All test are passing

+++ ComponentManager.cpp
@@ -1,9 +1,9 @@
 
 #include "ComponentManager.h"
 
-int ComponentManager::componentCount = 0;
+unsigned short ComponentManager::componentCount = 0;
 PackedArray<Sets*> ComponentManager::allComponentSets;
-std::map<std::string, int> ComponentManager::componentKeys;
+std::map<std::string, unsigned short> ComponentManager::componentKeys;
 std::vector<std::function<bool(unsigned short)>> ComponentManager::allEntityRemoves;
 //template <typename T>
 //ComponentSet<T> ComponentManager::CreateComponentSet()
